Try to move refinement types to their own modules
