+++ modules/api/src/main/scala/types.scala
@@ -1,20 +1,11 @@
-package fide
-package types
+package fide.spec
 
 import cats.syntax.all.*
 import fide.spec.*
-import io.github.iltotore.iron.*
-import io.github.iltotore.iron.constraint.all.*
 import smithy4s.*
+import fide.types.*
 
-opaque type Natural <: Int :| Positive = Int :| Positive
-
-object Natural extends RefinedTypeOps[Int, Positive, Natural]:
-  def fromString(value: String): Either[String, Natural] =
-    value.toIntOption.toRight(s"$value is not an int") >>= Natural.either
-
-  extension (self: Natural) def succ: Natural = Natural.applyUnsafe(self + 1)
-
+object providers:
   given RefinementProvider[PageFormat, String, Natural] =
     Refinement.drivenBy(Natural.fromString, _.toString)
